export-w3c-test-changes should allow the branch to be kept on failure https://bugs.webkit.org/show_bug.cgi?id=270348 rdar://problem/124269575
Reviewed by Sihui Liu.

This adds two arguments:

 * --no-clean, which toggles doing any cleanup, even on success
 * --clean-on-failure, which toggles cleaning up on failure

While we're at it, avoid needlessly running:

    git checkout master
    git reset --hard origin
    git checkout master

at the end of execution. We can just checkout once, and just leave
master as it is.

* Tools/Scripts/webkitpy/w3c/test_exporter.py:
(WebPlatformTestExporter.delete_local_branch):
(WebPlatformTestExporter.do_export):
(parse_args):
* Tools/Scripts/webkitpy/w3c/test_exporter_unittest.py:
(TestExporterTest.test_export):
(TestExporterTest.test_export_with_specific_branch):
(TestExporterTest):
(TestExporterTest.test_export_no_clean):

Canonical link: https://commits.webkit.org/286253@main

+++ Tools/Scripts/webkitpy/w3c/test_exporter.py
@@ -345,10 +345,13 @@ class WebPlatformTestExporter(object):
             _log.info('Error creating a pull request on github. Please ensure that the provided github token has the "public_repo" scope.')
         return pr_number
 
-    def delete_local_branch(self):
-        _log.info('Removing local branch ' + self._branch_name)
-        self._git.checkout('master')
-        self._git.delete_branch(self._branch_name)
+    def delete_local_branch(self, *, is_success=True):
+        if self._options.clean and (is_success or self._options.clean_on_failure):
+            _log.info('Removing local branch ' + self._branch_name)
+            self._git.checkout('master')
+            self._git.delete_branch(self._branch_name)
+        else:
+            _log.info('Keeping local branch ' + self._branch_name)
 
     def create_upload_remote_if_needed(self):
         if not self._wpt_fork_remote in self._git.remote([]):
@@ -365,29 +368,31 @@ class WebPlatformTestExporter(object):
         self.clean()
 
         if not self.create_branch_with_patch(git_patch_file):
-            _log.error("Cannot create web-platform-tests local branch from the patch")
-            self.delete_local_branch()
+            _log.error("Cannot create web-platform-tests local branch from the patch %r", git_patch_file)
+            self.delete_local_branch(is_success=False)
             return
 
-        if git_patch_file:
+        if git_patch_file and self.clean:
             self._filesystem.remove(git_patch_file)
 
         if self._options.use_linter:
             lint_errors = self._linter.lint()
             if lint_errors:
                 _log.error("The wpt linter detected %s linting error(s). Please address the above errors before attempting to export changes to the web-platform-test repository." % (lint_errors,))
-                self.delete_local_branch()
-                self.clean()
+                self.delete_local_branch(is_success=False)
                 return
 
         try:
             if self.push_to_wpt_fork():
                 if self._options.create_pull_request:
                     self.make_pull_request()
+        except Exception:
+            self.delete_local_branch(is_success=False)
+            raise
+        else:
+            self.delete_local_branch(is_success=True)
         finally:
-            self.delete_local_branch()
             _log.info("Finished")
-            self.clean()
 
 
 def parse_args(args):
@@ -426,6 +431,8 @@ def parse_args(args):
     parser.add_argument('-c', '--create-pr', dest='create_pull_request', action='store_true', default=False, help='create pull request to w3c web-platform-tests')
     parser.add_argument('--non-interactive', action='store_true', dest='non_interactive', default=False, help='Never prompt the user, fail as fast as possible.')
     parser.add_argument('--no-linter', action='store_false', dest='use_linter', default=True, help='Disable linter.')
+    parser.add_argument('--no-clean', action='store_false', dest='clean', help='Do not clean up.')
+    parser.add_argument('--clean-on-failure', action='store_true', dest='clean_on_failure', help='Do not clean up on failure.')
 
     options, args = parser.parse_known_args(args)
